requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

openai
azure-cognitiveservices-speech
python_dotenv
backoff
sounddevice
scipy
wavio
azure-storage-blob
azure-eventgrid
azure-functions%
pyodbc